Refactor: select.py including test_select.py
Move generic fixtures from test_select.py to conftest.py

Progress on 159
Progress on 158
